Feat/#18 exception handler

+++ src/main/java/com/nursy/nursy/domain/ward/WardService.java
@@ -6,6 +6,8 @@ import com.nursy.nursy.domain.ward.dto.WardUpdateRequestDto;
 import com.nursy.nursy.domain.user.User;
 import com.nursy.nursy.domain.user.UserRepository;
 import com.nursy.nursy.domain.wardSetting.WardSettingRepository;
+import com.nursy.nursy.global.exception.CustomException;
+import com.nursy.nursy.global.exception.GlobalErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
@@ -14,6 +16,8 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
+import static com.nursy.nursy.global.exception.GlobalErrorCode.USER_NOT_FOUND;
 
 @Service
 @RequiredArgsConstructor
@@ -27,15 +31,19 @@ public class WardService {
     @Transactional
     public Ward saveWard(Authentication authentication, WardAddRequestDto wardAddRequestDto) {
         Optional<User> userOptional = userRepository.findById(authentication.getName());//uuid
-        Ward ward = Ward.builder()
-                .wardName(wardAddRequestDto.getWardName())
-                .hospitalName(wardAddRequestDto.getHospitalName())
-                .location(wardAddRequestDto.getLocation())
-                .user(userOptional.get())
-                .build();
-        wardRepository.save(ward);
-        return ward;
-
+        if (userOptional.isPresent()) {
+            Ward ward = Ward.builder()
+                    .wardName(wardAddRequestDto.getWardName())
+                    .hospitalName(wardAddRequestDto.getHospitalName())
+                    .location(wardAddRequestDto.getLocation())
+                    .user(userOptional.get())
+                    .build();
+            wardRepository.save(ward);
+            return ward;
+        }
+        else{
+            throw new CustomException(GlobalErrorCode.USER_NOT_FOUND);
+        }
     }
     @Transactional
     public void deleteWard(Authentication authentication, Long wardId) {
@@ -43,12 +51,22 @@ public class WardService {
 
         if (userOptional.isPresent()) {
             User user = userOptional.get();
+            //유효성 검증 삭제 권한이 있는지 확인
+            boolean isManagingWard = user.getWards().stream()
+                    .anyMatch(ward -> ward.getWardId().equals(wardId));//wardId로 user찾기 vs userId로 ward찾기
+            if (!isManagingWard) {
+                throw new CustomException(GlobalErrorCode.WARD_ACCESS_DENIED);  // 커스텀 예외 처리
+            }
             wardRepository.deleteByUserIdAndWardId(user.getId(), wardId);
+        }
+        else{
+            throw new CustomException(GlobalErrorCode.USER_NOT_FOUND);
         }
     }
     @Transactional
     public void updateWard(Authentication authentication, WardUpdateRequestDto wardUpdateRequestDto) {
         Optional<User> userOptional = userRepository.findById(authentication.getName());//uuid
+
         Ward ward = Ward.builder()
                 .wardId(wardUpdateRequestDto.getWardId())
                 .wardName(wardUpdateRequestDto.getWardName())
@@ -78,9 +96,5 @@ public class WardService {
             // 예외나 빈 리스트 반환 가능
             return List.of();
         }
-    }
-    public Ward getWardById(Long wardId) {
-        Optional<Ward> wardOptional = wardRepository.findById(wardId);
-        return wardOptional.orElse(null);
     }
 }
